back to point where we log every html page recieved
after completely restarting w/ original facecat script, we're back to the point
of logging every html page recieved. that's in addition to logging every
function we enter. we need to 1) look at that logging thing again, and 2) find
some debugging platform. i'd like to be able to just look at the state of it.
still hate everything.